chore: fix previews
