+++ iHog/iHog/Localizable.xcstrings
@@ -287,9 +287,6 @@
     "Help" : {
 
     },
-    "Horizontal size: %@" : {
-
-    },
     "iHog Pro" : {
 
     },
@@ -532,9 +529,6 @@
 
     },
     "Upgrade to pro" : {
-
-    },
-    "Vertical size: %@" : {
 
     },
     "World" : {
